requirements.txt
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
future==0.18.2
idna==2.8
psycopg2-binary==2.8.4
pycparser==2.19
PySocks==1.7.1
python-telegram-bot==13.4.1
PyYAML==5.3
requests==2.22.0
six==1.13.0
tornado==6.0.3
urllib3==1.25.7
yoyo-migrations==6.1.0
german_nouns==1.2.1
